refactor the server, move tcp listener to a manager
