+++ tunneld-pkg/src/lib.rs
@@ -1,2 +1,3 @@
 pub mod io;
 pub mod event;
+pub mod util;
